kinda just seems like this should be a ListPageView
