+++ src/modules/cmdpal/Microsoft.CmdPal.UI/ExtViews/ParametersPage.xaml
@@ -3,72 +3,14 @@
     x:Class="Microsoft.CmdPal.UI.ParametersPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:Interactions="using:Microsoft.Xaml.Interactions.Core"
-    xmlns:Interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:cmdPalControls="using:Microsoft.CmdPal.UI.Controls"
     xmlns:cmdpalUI="using:Microsoft.CmdPal.UI"
     xmlns:controls="using:CommunityToolkit.WinUI.Controls"
-    xmlns:converters="using:CommunityToolkit.WinUI.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:help="using:Microsoft.CmdPal.UI.Helpers"
-    xmlns:local="using:Microsoft.CmdPal.UI"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:viewModels="using:Microsoft.CmdPal.UI.ViewModels"
     Background="Transparent"
     mc:Ignorable="d">
-
-    <Page.Resources>
-        <ResourceDictionary>
-            <DataTemplate x:Key="ParameterListItemTemplate" x:DataType="viewModels:ListItemViewModel">
-                <Grid AutomationProperties.Name="{x:Bind Title, Mode=OneWay}" ColumnSpacing="12">
-                    <Grid.ColumnDefinitions>
-                        <ColumnDefinition Width="28" />
-                        <ColumnDefinition Width="*" />
-                    </Grid.ColumnDefinitions>
-
-                    <cmdPalControls:IconBox
-                        Grid.Column="0"
-                        Width="20"
-                        Height="20"
-                        AutomationProperties.AccessibilityView="Raw"
-                        Foreground="{ThemeResource TextFillColorSecondaryBrush}"
-                        SourceKey="{x:Bind Icon, Mode=OneWay}"
-                        SourceRequested="{x:Bind help:IconProvider.SourceRequested20}" />
-
-                    <Grid Grid.Column="1" ColumnSpacing="12">
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="Auto" />
-                            <ColumnDefinition Width="*" />
-                        </Grid.ColumnDefinitions>
-
-                        <TextBlock
-                            Grid.Column="0"
-                            VerticalAlignment="Center"
-                            CharacterSpacing="12"
-                            FontSize="14"
-                            MaxLines="1"
-                            Style="{StaticResource CaptionTextBlockStyle}"
-                            Text="{x:Bind Title, Mode=OneWay}"
-                            TextTrimming="CharacterEllipsis"
-                            TextWrapping="NoWrap" />
-
-                        <TextBlock
-                            Grid.Column="1"
-                            VerticalAlignment="Center"
-                            CharacterSpacing="12"
-                            FontSize="14"
-                            Foreground="{ThemeResource TextFillColorSecondaryBrush}"
-                            MaxLines="1"
-                            Style="{StaticResource CaptionTextBlockStyle}"
-                            Text="{x:Bind Subtitle, Mode=OneWay}"
-                            TextTrimming="CharacterEllipsis"
-                            TextWrapping="NoWrap"
-                            Visibility="{x:Bind ShowSubtitle, Mode=OneWay}" />
-                    </Grid>
-                </Grid>
-            </DataTemplate>
-        </ResourceDictionary>
-    </Page.Resources>
 
     <Grid>
         <controls:SwitchPresenter
@@ -76,18 +18,11 @@
             TargetType="x:Boolean"
             Value="{x:Bind ViewModel.HasActiveList, Mode=OneWay}">
             <controls:Case Value="True">
-                <!--  Show list items from the active list parameter  -->
-                <ListView
-                    x:Name="ParamItemsList"
-                    Padding="0,2,0,0"
-                    IsItemClickEnabled="True"
-                    ItemClick="ParamItems_ItemClick"
-                    ItemTemplate="{StaticResource ParameterListItemTemplate}"
-                    ItemsSource="{x:Bind ViewModel.ActiveListViewModel.FilteredItems, Mode=OneWay}">
-                    <ListView.ItemContainerTransitions>
-                        <TransitionCollection />
-                    </ListView.ItemContainerTransitions>
-                </ListView>
+                <!--
+                    Show list items from the active list parameter using the same
+                    list view used by ListPage so behavior stays in one place.
+                -->
+                <cmdpalUI:ListItemsView ViewModel="{x:Bind ViewModel.ActiveListViewModel, Mode=OneWay}" />
             </controls:Case>
             <controls:Case Value="False">
                 <controls:SwitchPresenter
