+++ src/modules/cmdpal/Microsoft.CmdPal.UI/ExtViews/ParametersPage.xaml.cs
@@ -2,13 +2,9 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using CommunityToolkit.Mvvm.Messaging;
-using ManagedCommon;
 using Microsoft.CmdPal.Common;
 using Microsoft.CmdPal.UI.ViewModels;
 using Microsoft.CmdPal.UI.ViewModels.Commands;
-using Microsoft.CmdPal.UI.ViewModels.Messages;
-using Microsoft.UI.Dispatching;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Navigation;
@@ -16,17 +12,12 @@ using Microsoft.UI.Xaml.Navigation;
 namespace Microsoft.CmdPal.UI;
 
 /// <summary>
-/// An empty page that can be used on its own or navigated to within a Frame.
+/// Hosts a parameter run, optionally embedding a <see cref="ListItemsView"/> when
+/// a list parameter is active. List rendering, selection, and keyboard navigation
+/// are handled by the embedded <see cref="ListItemsView"/>.
 /// </summary>
-public sealed partial class ParametersPage : Page,
-    IRecipient<NavigateNextCommand>,
-    IRecipient<NavigatePreviousCommand>,
-    IRecipient<NavigatePageDownCommand>,
-    IRecipient<NavigatePageUpCommand>,
-    IRecipient<ActivateSelectedListItemMessage>
+public sealed partial class ParametersPage : Page
 {
-    private readonly DispatcherQueue _queue = DispatcherQueue.GetForCurrentThread();
-
     public ParametersPageViewModel? ViewModel
     {
         get => (ParametersPageViewModel?)GetValue(ViewModelProperty);
@@ -40,20 +31,6 @@ public sealed partial class ParametersPage : Page,
     public ParametersPage()
     {
         this.InitializeComponent();
-        this.Unloaded += OnUnloaded;
-
-        WeakReferenceMessenger.Default.Register<NavigateNextCommand>(this);
-        WeakReferenceMessenger.Default.Register<NavigatePreviousCommand>(this);
-        WeakReferenceMessenger.Default.Register<NavigatePageDownCommand>(this);
-        WeakReferenceMessenger.Default.Register<NavigatePageUpCommand>(this);
-        WeakReferenceMessenger.Default.Register<ActivateSelectedListItemMessage>(this);
-    }
-
-    private void OnUnloaded(object sender, RoutedEventArgs e)
-    {
-        // Unhook from everything to ensure nothing can reach us
-        // between this point and our complete and utter destruction.
-        WeakReferenceMessenger.Default.UnregisterAll(this);
     }
 
     protected override void OnNavigatedTo(NavigationEventArgs e)
@@ -83,104 +60,9 @@ public sealed partial class ParametersPage : Page,
 
     private static void OnViewModelChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
     {
-        if (d is ParametersPage @this)
+        if (d is ParametersPage && e.NewValue is null)
         {
-            if (e.OldValue is ParametersPageViewModel old)
-            {
-                old.PropertyChanged -= @this.ViewModel_PropertyChanged;
-            }
-
-            if (e.NewValue is ParametersPageViewModel page)
-            {
-                page.PropertyChanged += @this.ViewModel_PropertyChanged;
-            }
-            else if (e.NewValue is null)
-            {
-                CoreLogger.LogDebug("cleared view model");
-            }
+            CoreLogger.LogDebug("cleared view model");
         }
-    }
-
-    private void ViewModel_PropertyChanged(object? sender, System.ComponentModel.PropertyChangedEventArgs e)
-    {
-        var prop = e.PropertyName;
-        if (prop == nameof(ViewModel.ShowCommand))
-        {
-            CoreLogger.LogDebug($"ViewModel.ShowCommand {ViewModel?.ShowCommand}");
-        }
-        else if (prop == nameof(ViewModel.ActiveListViewModel))
-        {
-            if (ViewModel?.HasActiveList == true)
-            {
-                SelectFirstItem();
-            }
-        }
-    }
-
-    private void ParamItems_ItemClick(object sender, ItemClickEventArgs e)
-    {
-        if (e.ClickedItem is ListItemViewModel item)
-        {
-            ViewModel?.ActiveListViewModel?.InvokeItemCommand.Execute(item);
-        }
-    }
-
-    public void Receive(NavigateNextCommand message) => NavigateList(1);
-
-    public void Receive(NavigatePreviousCommand message) => NavigateList(-1);
-
-    public void Receive(NavigatePageDownCommand message) => NavigateList(10);
-
-    public void Receive(NavigatePageUpCommand message) => NavigateList(-10);
-
-    public void Receive(ActivateSelectedListItemMessage message)
-    {
-        if (ViewModel?.HasActiveList != true)
-        {
-            return;
-        }
-
-        if (ParamItemsList.SelectedItem is ListItemViewModel item)
-        {
-            ViewModel.ActiveListViewModel?.InvokeItemCommand.Execute(item);
-        }
-        else if (ViewModel.ActiveListViewModel?.FilteredItems.Count > 0 &&
-                 ViewModel.ActiveListViewModel.FilteredItems[0] is ListItemViewModel firstItem)
-        {
-            ViewModel.ActiveListViewModel.InvokeItemCommand.Execute(firstItem);
-        }
-    }
-
-    private void NavigateList(int delta)
-    {
-        if (ViewModel?.HasActiveList != true)
-        {
-            return;
-        }
-
-        var list = ParamItemsList;
-        var count = list.Items.Count;
-        if (count == 0)
-        {
-            return;
-        }
-
-        var current = list.SelectedIndex;
-        var target = Math.Clamp(current + delta, 0, count - 1);
-        list.SelectedIndex = target;
-        list.ScrollIntoView(list.SelectedItem);
-    }
-
-    public void SelectFirstItem()
-    {
-        // Use TryEnqueue so the ListView has had time to populate from the binding
-        _queue.TryEnqueue(() =>
-        {
-            if (ParamItemsList.Items.Count > 0)
-            {
-                ParamItemsList.SelectedIndex = 0;
-                ParamItemsList.ScrollIntoView(ParamItemsList.SelectedItem);
-            }
-        });
     }
 }
